defined a new type of Characteristic for quadratic spline interpolation (benefit: ready for file I/O), moved the helper function to pp.control.util.auxiliary, updated the jupyter notebook
