+++ pandapower/control/__init__.py
@@ -12,7 +12,7 @@ from pandapower.control.controller.trafo_control import TrafoController
 # --- Other ---
 from pandapower.control.run_control import *
 from pandapower.control.run_control import ControllerNotConverged
-from pandapower.control.util.characteristic import Characteristic
-from pandapower.control.util.auxiliary import get_controller_index
+from pandapower.control.util.characteristic import Characteristic, SplineCharacteristic
+from pandapower.control.util.auxiliary import get_controller_index, plot_characteristic, create_trafo_characteristics
 from pandapower.control.util.diagnostic import control_diagnostic
