+++ pandapower/control/util/auxiliary.py
@@ -14,6 +14,8 @@ import pandas as pd
 from pandas import Int64Index
 
 from pandapower.toolbox import ensure_iterability
+from .characteristic import SplineCharacteristic
+
 try:
     import matplotlib.pyplot as plt
     MATPLOTLIB_INSTALLED = True
@@ -208,3 +210,25 @@ def plot_characteristic(characteristic, start, stop, num=20):
         plt.plot(x, y, marker='x')
     else:
         logger.info("matplotlib not installed. y-values: %s" % y)
+
+
+def create_trafo_characteristics(net, trafotable, trafo_index, vk_percent_x, vk_percent_y, vkr_percent_x, vkr_percent_y):
+    if len(trafo_index) != len(vk_percent_x) or len(trafo_index) != len(vkr_percent_x) or \
+            len(trafo_index) != len(vk_percent_y) or len(trafo_index) != len(vkr_percent_y):
+        raise UserWarning("the lengths of the trafo index and points do not match!")
+
+    if 'tap_dependent_impedance' not in net[trafotable]:
+        net[trafotable]['tap_dependent_impedance'] = False
+
+    net[trafotable].loc[trafo_index, 'tap_dependent_impedance'] = True
+
+    for col in ('vk_percent_characteristic', 'vkr_percent_characteristic'):
+        if col not in net[trafotable]:
+            net[trafotable][col] = pd.Series(index=net[trafotable].index, dtype=object, data=None)
+        elif net[trafotable][col].dtype != object:
+            net[trafotable][col] = net[trafotable][col].astype(object)
+
+    for tid, vk_x, vk_y, vkr_x, vkr_y in zip(trafo_index, vk_percent_x, vk_percent_y, vkr_percent_x, vkr_percent_y):
+        net[trafotable].at[tid, 'vk_percent_characteristic'] = SplineCharacteristic(vk_x, vk_y)
+        net[trafotable].at[tid, 'vkr_percent_characteristic'] = SplineCharacteristic(vkr_x, vkr_y)
+
